move to interface
